==> 01-box-office-mojo-scraping.ipynb <==
# -*- coding: utf-8 -*-

from __future__ import division, unicode_literals

import pandas as pd
import numpy as np

import dateutil
import datetime
import time
from time import sleep

from bs4 import BeautifulSoup
import requests

import re
import html5lib

import pickle

base_url = 'http://www.boxofficemojo.com'

base_year_url = base_url + '/yearly/'

columns = ['Title', 'DomesticTotalGross', 'Distributor', 'Runtime', 'MPAARating', 'ReleaseDate', 'Genre', 
           'ProductionBudget','Director']

def make_soup(url):
    response = requests.get(url)
    page = BeautifulSoup(response.text, 'html5lib')
    return page.find(id = 'main')

def get_year_url(year):
    return base_year_url + 'chart/?yr=' + str(year) + '&p=.htm'

## DEPRECIATED
# Get all the links to year pages
def get_years(url):
    soup = make_soup(url)
    return [url + year['href'] for year in soup.find_all('a', href = re.compile('chart/\?yr'))]

# From each year page, get the links to each page
def get_year_pages(url):
    soup = make_soup(url)
    return list(set([base_url + year_page['href'] for year_page in soup.find_all('a', href = re.compile('chart/\?page'))]))

# From each page, get the links to each movie page
def get_movie_pages(url):
    soup = make_soup(url)
    return [(movie_page.text, base_url + movie_page['href']) for movie_page in soup.find_all('a', href = re.compile('/movies/'))]

# DEPRECIATED
# From each page, get the name of each movie
def get_movie_names(url):
    soup = make_soup(url)
    return [movie_name.text for movie_name in soup.find_all('a', href = re.compile('/movies/'))]

# Define helper "getter" function 
def get_movie_value(soup, field_name):
    obj = soup.find(text=re.compile(field_name))
    if not obj:
        return None
    next_sibling = obj.findNextSibling()
    next_ = obj.findNext()
    if next_sibling:
        return next_sibling.text
    elif next_:
        return next_.text
    else:
        return None

# Including transformation functions
def to_date(datestring):
    try:
        date = dateutil.parser.parse(datestring)
        return date
    except:
        return datestring
    
def money_to_int(moneystring):
    num = [('million', 6), ('thousand', 3)]
    moneystring = moneystring.replace('$', '').replace(',', '').replace(' ','').replace('(Estimate)','').lower()
    for n in num:
        if n[0] in moneystring:
            moneystring = moneystring.replace(n[0], '')
            return int(moneystring) * np.power(10,n[1])
    else:
        return int(moneystring)

def runtime_to_minutes(runtimestring):
    runtime = runtimestring.split()
    try:
        minutes = int(runtime[0])*60 + int(runtime[2])
        return minutes
    except:
        return None

# For each movie page, get the following items

def get_movie_info(url, d):
    soup = make_soup(url) 
    
    # Domestic Total Gross
    try:
        d['DomesticTotalGross'] = money_to_int(get_movie_value(soup, 'Domestic Total'))
    except:
        pass
    
    # Distributor
    try:
        d['Distributor'] = get_movie_value(soup, 'Distributor')
    except: 
        pass
    
    # Runtime
    try:
        d['Runtime'] = runtime_to_minutes(get_movie_value(soup, 'Runtime'))
    except:
        pass
    
    # MPAA Rating
    try: 
        d['MPAARating'] = get_movie_value(soup, 'MPAA Rating')
    except:
        pass
    
    # Release Date
    try:
        d['ReleaseDate'] = to_date(get_movie_value(soup, 'Release Date'))
    except:
        pass
        
    # Genre    
    try: 
        d['Genre'] = get_movie_value(soup, 'Genre[^s]')
    except:
        pass
        
    # Production Budget
    try:
        d['ProductionBudget'] = money_to_int(get_movie_value(soup, 'Production Budget'))
    except:
        pass
    
    # Director
    try: 
        d['Director'] = get_movie_value(soup, 'Director')
    except:
        pass

# timing function
def timefunc(f):
    def f_timer(*args, **kwargs):
        start = time.time()
        result = f(*args, **kwargs)
        end = time.time()
        print f.__name__, 'took', end - start, 'seconds'
        return result
    return f_timer

@timefunc
def scrape_all_year_pages(year_url):
    df = scrape_year_page(year_url)
    year_pages = get_year_pages(year_url)
    for year_page in year_pages:
        df = df.append(scrape_year_page(year_page), ignore_index=True)
    return df

@timefunc  
def scrape_year_page(year_page_url):
    df = pd.DataFrame(columns = columns)
    named_pages = get_movie_pages(year_page_url)
    for name, movie_page in named_pages:
        # Initialize an empty dictionary
        dic = dict.fromkeys(columns)
        # Set title to movie name
        dic['Title'] = name
        # Collect additional movie information
        get_movie_info(movie_page, dic)
        # Append to dataframe
        df = df.append(dic, ignore_index=True)
        sleep(.05)     
    return df

@timefunc
# For scraping an individual year and saving to pickle
def scrape_year(year):
    year_url = get_year_url(year)
    df = scrape_all_year_pages(year_url)
    df.to_pickle('data/box_office_mojo_' + str(year) + '.pkl')

@timefunc
# For scraping a range of years
def scrape_years(start, stop):
    years = range(stop, start-1, -1)
    for year in years:
        scrape_year(year)

## Scraping Begins

# this was an unintentionally manual process - originally saved each year as an individual variable, 
# but with pickling can run in a loop without needing to store as variables

year_2016 = scrape_all_year_pages('http://www.boxofficemojo.com/yearly/chart/?yr=2016&p=.htm')

year_2016.to_pickle('data/box_office_mojo_2016.pkl')

year_2015 = scrape_all_year_pages('http://www.boxofficemojo.com/yearly/chart/?yr=2015&p=.htm')

year_2015.to_pickle('data/box_office_mojo_2015.pkl')

year_2014 = scrape_years(base_year_url, 2014, 2014)

year_2014 = year_2014[0]

year_2013 = scrape_years(base_year_url, 2013, 2013)

year_2013 = year_2013[0]

year_2012 = scrape_year(base_year_url, 2012)

year_2011 = scrape_year(base_year_url, 2011)

year_2010 = scrape_year(base_year_url, 2010)

year_2009 = scrape_year(base_year_url, 2009)

year_2008 = scrape_year(base_year_url, 2008)

year_2007 = scrape_year(base_year_url, 2007)
year_2006 = scrape_year(base_year_url, 2006)
year_2005 = scrape_year(base_year_url, 2005)
year_2004 = scrape_year(base_year_url, 2004)

year_2003 = scrape_year(base_year_url, 2003)
year_2002 = scrape_year(base_year_url, 2002)
year_2001 = scrape_year(base_year_url, 2001)
year_2000 = scrape_year(base_year_url, 2000)

year_1999 = scrape_year(base_year_url, 1999)
year_1998 = scrape_year(base_year_url, 1998)
year_1997 = scrape_year(base_year_url, 1997)
year_1996 = scrape_year(base_year_url, 1996)

year_1995 = scrape_year(base_year_url, 1995)
year_1994 = scrape_year(base_year_url, 1994)
year_1993 = scrape_year(base_year_url, 1993)
year_1992 = scrape_year(base_year_url, 1992)
year_1991 = scrape_year(base_year_url, 1991)
year_1990 = scrape_year(base_year_url, 1990)

scrape_years(1980,1986)

## Test Functions

# get_years('http://www.boxofficemojo.com/yearly/')
# get_year_pages('http://www.boxofficemojo.com/yearly/chart/?yr=2016&p=.htm')
# get_movie_pages('http://www.boxofficemojo.com/yearly/chart/?page=2&view=releasedate&view2=domestic&yr=2016&p=.htm')
# get_movie_names('http://www.boxofficemojo.com/yearly/chart/?page=2&view=releasedate&view2=domestic&yr=2016&p=.htm')
# l2016 = scrape_year_page('http://www.boxofficemojo.com/yearly/chart/?yr=2016&p=.htm')

## Test Load Pickle

pd.read_pickle('data/box_office_mojo_2013.pkl')
