==> 04-omdb-querying.ipynb <==
# -*- coding: utf-8 -*-

from __future__ import division, unicode_literals

import pandas as pd
import numpy as np

import time

%matplotlib inline

import pickle
import omdb

def timefunc(f):
    def f_timer(*args, **kwargs):
        start = time.time()
        result = f(*args, **kwargs)
        end = time.time()
        print f.__name__, 'took', end - start, 'seconds'
        return result
    return f_timer

omdb.set_default('tomatoes', True)
omdb.set_default('plot', 'short')

def get_imdbID_list(pickle):
    return list(pd.read_pickle('data/imdbid.pkl'))

@timefunc
def get_omdb_data(ids):
    l = []
    for i in ids:
        try:
            l.append(dict(omdb.imdbid(i, tomatoes=True)))
        except:
            print "Exceptions with ID %s" %i
    return pd.DataFrame(l)

imdbIDs = get_imdbID_list('data/imdbid.pkl')

# Data acquisition
omdb_data = get_omdb_data(imdbIDs)

omdb_data.to_pickle('data/omdb.pkl')

## Load Pickle Point

omdb_data = pd.read_pickle('data/omdb.pkl')

omdb_data.columns

omdb_data['awards'].unique()
