==> 05-data-cleaning.ipynb <==
## Loading and Cleaning  Data

# -*- coding: utf-8 -*-

from __future__ import division, unicode_literals

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

%matplotlib inline

import dateutil
import datetime
import time

import re

import pickle

from genderizer.genderizer import Genderizer

# from fuzzywuzzy import fuzz
# from fuzzywuzzy import process

def load_pickles(start, stop):
    df = pd.DataFrame([])
    for i in range(stop, start-1, -1):
        df = df.append(pd.read_pickle('data/box_office_mojo_'+str(i)+'.pkl'), ignore_index=True)
    return df

## Load Pickle Point

# Load pickle data
box_office_mojo = load_pickles(1980, 2016) # 14269 records
omdb = pd.read_pickle('data/omdb.pkl') # 2000 records
bechdel = pd.read_pickle('data/bechdel.pkl') # 5420 records
polygraph = pd.read_pickle('data/polygraph.pkl') # 1944 records

# Initial drop of unnecessary/duplicate columns
bechdel.drop(['title', 'year'], axis=1, inplace=True)
polygraph.drop(['script_id', 'title', 'year', 'gross'], axis=1, inplace=True)
omdb.drop(['box_office', 'dvd', 'plot', 'poster', 'response', 'tomato_consensus', 'tomato_fresh', 
           'tomato_image', 'tomato_rotten', 'type', 'website'], axis=1, inplace=True)
box_office_mojo.drop('Director', axis=1, inplace=True)

# Merging dataframes by unique IMDB IDs
polygraph_bechdel = pd.merge(polygraph, bechdel, on = 'imdb_id') # 1244 records
omdb_polygraph_bechdel = pd.merge(omdb, polygraph_bechdel, on = 'imdb_id') # 1244 records

polygraph.describe()

# Pickling for posterity
omdb_polygraph_bechdel.to_pickle('data/omdb_polygraph_bechdel.pkl')

# Including transformation functions
def to_date(datestring):
    try:
        date = dateutil.parser.parse(datestring)
        return date
    except:
        return datestring
    
def money_to_int(moneystring):
    num = [('million', 6), ('thousand', 3)]
    moneystring = moneystring.replace('$', '').replace(',', '').replace(' ','').replace('(Estimate)','').lower()
    for n in num:
        if n[0] in moneystring:
            moneystring = moneystring.replace(n[0], '')
            return int(moneystring) * np.power(10,n[1])
    else:
        return int(moneystring)

## Fuzzy Wuzzy Match - Not Used

# Using year for joins
# box_office_mojo['year'] = box_office_mojo['ReleaseDate'].dt.year
# omdb_polygraph_bechdel['year'] = omdb_polygraph_bechdel['year'].apply(float)
# merged = pd.merge(box_office_mojo, omdb_polygraph_bechdel, on = 'year', suffixes = ['_1', '_2'])

# merged['released'] = merged['released'].apply(to_date)
# merged['box-title-release'] = merged['Title'] + merged['ReleaseDate'].apply(str)
# merged['omdb-title-release'] = merged['title'] + merged['released'].apply(str)

# 1,224 - 1,117 records lost to exact merge
# 574,038 when merging on year

# def fuzztitle(row):
#     try:
#         return fuzz.ratio(row['box-title-release'], row['omdb-title-release'])
#     except:
#         return 0
    
# merged.loc[:, 'ratio'] = merged.apply(fuzztitle, axis = 1)
# merged[(merged['ratio']>80) & (merged['ratio']<100)].loc[:,['Title','title','ReleaseDate','released','ratio']]

## Final merge

box_office_mojo = box_office_mojo.rename(columns={'Title':'title'})
merged = pd.merge(box_office_mojo, omdb_polygraph_bechdel, on = 'title') # 1117 records

merged.head()

# Second drop of unnecessary/duplicate columns and renaming of columns

# Duplicates: Genres (Multiple), Runtime, MPAA Rating, Release Date
# Unnecessary (for now): Distributor, Producer

# Drop - Genre, runtime, rated, released, Distributor, production

merged.drop(['Genre', 'runtime', 'rated', 'released', 'Distributor', 'production'], axis=1, inplace=True)
merged = merged.rename(columns = {'DomesticTotalGross': 'domestic_total_gross', 'Runtime': 'runtime', 'MPAARating': 'mpaa', 
                        'ReleaseDate': 'release_date', 'ProductionBudget': 'production_budget'})

# Pickle and retrieve dataframe
merged.to_pickle('data/merged.pkl')

## Load Pickle Point

merged = pd.read_pickle('data/merged.pkl')

## Mutating and Dummies

# Helper function
def find_chosen_one(i, s):
    return int(i in s)

# Create # wins and # nominations
def return_wins(s):
    wins_regex = re.compile('((?<=[Ww]on )[0-9]|[0-9]*(?= [Ww]ins*))')
    wins = 0
    for i in re.findall(wins_regex, s):
        try:
            wins += int(i)
        except:
            pass
    return wins

def return_nominations(s):
    nominations_regex = re.compile('((?<=[Nn]ominated for )[0-9]|[0-9]*(?= [Nn]ominations*))')
    nominations = 0
    for i in re.findall(nominations_regex, s):
        try:
            nominations += int(i)
        except:
            pass
    return nominations

def create_awards(df):
    df['wins'] = df['awards'].apply(return_wins)
    df['nominations'] = df['awards'].apply(return_nominations)
    return df

## Getting Genderizations

# Returns the gender of a name based on the first name (defined as first word in name)
def return_gender(name):
    return Genderizer.detect(firstName = name.split(' ')[0])

# Returns the ratio of female to male agents (0 if all male, 1 if all female, fractional if mix)
# If gender is unknown, assume male
def genderize(s):
    l = []
    for i in s.split(', '):
        l.append(return_gender(i))
    return l.count('female')/len(l)

# Creates a genderization column in a df
def create_genderization(df, column):
    colname = column + '_genderization'
    df[colname] = df[column].apply(genderize)
    return df

## Getting Binaries

# Creates a dummy variable for whether the 'chosen one' (e.g., "USA", "English") is present or not
# Used for identifying foreign/non-English language films
def create_binaries(df, column, chosen_one):
    colname = column + '_' + chosen_one.lower()
    df[colname] = df[column].apply(lambda x: find_chosen_one(chosen_one, x)) 
    return df

## Getting Dummies

# Turn list-like, categorical columns into dummy variables
# Getting list of unique categorical levels
def find_unique(df, column):
    l = [i for i in df.loc[:,column]]
    l = reduce(lambda x, y: x + ', ' + y, l).split(',')
    return list(set(map(lambda x: x.strip(), l)))

# Creating dummy variables from each level (for multiple categories in single cell)
def create_dummies(df, column):
    l = find_unique(df, column)
    for i in l:
        colname = column + '_' + i.replace(' ','_').lower()
        df[colname] = df[column].apply(lambda x: find_chosen_one(i, x))
    return df

# Ready to edit dataframe for variables!
final = merged.copy()

# Create Awards - 'wins', 'nominations', drop awards
final = create_awards(final)
final.drop(['awards'], axis=1, inplace=True)

# Person Categories - Genderize - 'writer', 'actors', 'director', drop originals and also f, m
final = create_genderization(final, 'writer')
final = create_genderization(final, 'actors')
final = create_genderization(final, 'director')
final.drop(['writer', 'actors', 'director', 'f', 'm'], axis=1, inplace=True)

# Binary Categories - Create Binaries - 'country_usa', 'language_english', drop originals
final = create_binaries(final, 'country', 'USA')
final = create_binaries(final, 'language', 'English')
final.drop(['country','language'], axis=1, inplace=True)

# Unique Categories - Create Dummies - 'genre_...', 'mpaa_...' - possibly include producer & distributor in here
final = create_dummies(final, 'genre')
final = pd.get_dummies(final, columns = ['mpaa'])
final.drop(['genre'], axis=1, inplace=True)

# Final cleanup
final[final.duplicated(subset = 'imdb_id', keep=False)].loc[:,['imdb_id', 'title','release_date', 'year']]

# Removing duplicates manually (just a handful)
final.drop([28, 48, 49, 275, 591, 1082], inplace=True)

# Changing datatypes
final['release_date'] = final['release_date'].apply(to_date).apply(datetime.date.toordinal)
final['imdb_votes'] = final['imdb_votes'].apply(money_to_int)

def change_type(df):
    objects = df.select_dtypes(include=['object']).drop(['title','imdb_id'], axis = 1).columns
    for i in objects:
        try:
            df[i] = df[i].replace('N/A',np.nan)
            df[i] = pd.to_numeric(df[i], errors = 'raise')
        except:
            print i
    return df

# Check dtype
# for i in final3.columns:
#     print i, final3[i].dtype

final2 = change_type(final)

# dropping movies with no production budget listed
final3 = final2[final2['production_budget'].notnull()].copy()

# Replacing NA values with mean
nas = final3.isnull().sum()
nas_index = nas[nas > 0].index

def replace_nas(df, columns):
    for i in df[columns]:
        final[i] = final[i].fillna(value = final[i].mean())
    return df

final4 = replace_nas(final3, nas_index)

# Identifying outliers
final4['gross_to_budget'] = final4['domestic_total_gross'] / final4['production_budget']
final4.sort_values(by = 'gross_to_budget', ascending = False)

# Excluding movies where gross to budget ratio is high (skewed)
final5 = final4[final4['gross_to_budget']<10].copy()
final5.drop('gross_to_budget', axis=1, inplace=True)

movies = final5.set_index(['title', 'imdb_id'])

movies.to_pickle('data/movies.pkl')

## Load Pickle Point

movies = pd.read_pickle('data/movies.pkl')

bechdel.describe()

omdb_polygraph_bechdel.describe()

((omdb_polygraph_bechdel['f']) / (omdb_polygraph_bechdel['total_lines'])).median()
